Styling improvements.
