+++ packages/running/src/index.tsx
@@ -61,6 +61,11 @@ const ITEM_LABEL_CLASS = 'jp-RunningSessions-itemLabel';
  */
 const SHUTDOWN_BUTTON_CLASS = 'jp-RunningSessions-itemShutdown';
 
+/**
+ * The class name added to a running session item shutdown button.
+ */
+const SHUTDOWN_ALL_BUTTON_CLASS = 'jp-RunningSessions-shutdownAll';
+
 /* tslint:disable */
 /**
  * The running sessions token.
@@ -117,11 +122,10 @@ export class RunningSessionManagers implements IRunningSessionManagers {
 
 function Item(props: {
   runningItem: IRunningSessions.IRunningItem;
+  shutdownLabel: string;
   translator?: ITranslator;
 }) {
   const { runningItem } = props;
-  const translator = props.translator || nullTranslator;
-  const trans = translator.load('jupyterlab');
   const icon = runningItem.icon();
 
   return (
@@ -134,18 +138,20 @@ function Item(props: {
       >
         {runningItem.label()}
       </span>
-      <button
-        className={`${SHUTDOWN_BUTTON_CLASS} jp-mod-styled`}
+      <ToolbarButtonComponent
+        className={SHUTDOWN_BUTTON_CLASS}
+        icon={closeIcon}
         onClick={() => runningItem.shutdown()}
-      >
-        {trans.__('SHUT DOWN')}
-      </button>
+        tooltip={props.shutdownLabel}
+      />
     </li>
   );
 }
 
 function ListView(props: {
   runningItems: IRunningSessions.IRunningItem[];
+  shutdownLabel: string;
+  shutdownAllLabel: string;
   translator?: ITranslator;
 }) {
   return (
@@ -154,6 +160,7 @@ function ListView(props: {
         <Item
           key={i}
           runningItem={item}
+          shutdownLabel={props.shutdownLabel}
           translator={props.translator || nullTranslator}
         />
       ))}
@@ -163,6 +170,8 @@ function ListView(props: {
 
 function List(props: {
   manager: IRunningSessions.IManager;
+  shutdownLabel: string;
+  shutdownAllLabel: string;
   translator?: ITranslator;
 }) {
   return (
@@ -170,6 +179,8 @@ function List(props: {
       {() => (
         <ListView
           runningItems={props.manager.running()}
+          shutdownLabel={props.shutdownLabel}
+          shutdownAllLabel={props.shutdownAllLabel}
           translator={props.translator || nullTranslator}
         />
       )}
@@ -189,13 +200,20 @@ function Section(props: {
 }) {
   const translator = props.translator || nullTranslator;
   const trans = translator.load('jupyterlab');
+  const shutdownLabel = props.manager.shutdownLabel || trans.__('Shut Down');
+  const shutdownAllLabel =
+    props.manager.shutdownAllLabel || trans.__('Shut Down All');
+  const shutdownTitle = `${shutdownAllLabel}?`;
+  const shutdownAllConfirmationText =
+    props.manager.shutdownAllConfirmationText ||
+    `${shutdownAllLabel} ${props.manager.name}`;
   function onShutdown() {
     void showDialog({
-      title: trans.__('Shut Down All?'),
-      body: `${props.manager.name}`,
+      title: shutdownTitle,
+      body: shutdownAllConfirmationText,
       buttons: [
         Dialog.cancelButton({ label: trans.__('Cancel') }),
-        Dialog.warnButton({ label: trans.__('SHUT DOWN') })
+        Dialog.warnButton({ label: shutdownAllLabel })
       ]
     }).then(result => {
       if (result.button.accept) {
@@ -209,15 +227,21 @@ function Section(props: {
       <>
         <header className={SECTION_HEADER_CLASS}>
           <h2>{props.manager.name}</h2>
-          <ToolbarButtonComponent
-            icon={closeIcon}
+          <button
+            className={`${SHUTDOWN_ALL_BUTTON_CLASS} jp-mod-styled`}
             onClick={onShutdown}
-            tooltip={trans.__('Shut Down All…')}
-          />
+          >
+            {shutdownAllLabel}
+          </button>
         </header>
 
         <div className={CONTAINER_CLASS}>
-          <List manager={props.manager} translator={props.translator} />
+          <List
+            manager={props.manager}
+            shutdownLabel={shutdownLabel}
+            shutdownAllLabel={shutdownAllLabel}
+            translator={props.translator}
+          />
         </div>
       </>
     </div>
@@ -299,6 +323,10 @@ export namespace IRunningSessions {
     refreshRunning(): void;
     // A signal that should be emitted when the item list has changed.
     runningChanged: ISignal<any, any>;
+    // A string used to describe the shutdown action.
+    shutdownLabel?: string;
+    shutdownAllLabel?: string;
+    shutdownAllConfirmationText?: string;
   }
 
   /**
